requirements.txt
Flask==3.1.0
Flask-WTF==1.2.2
cryptography==44.0.2
openpyxl==3.1.5
python-dotenv==1.0.1
requests==2.32.3
WTForms==3.2.1